pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kano-agent"
version = "1.0.0"
description = "Kano Model Feature Prioritization Agent using OpenAI Agents SDK"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.5",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["classification*", "analysis*", "report*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
